requirements.txt
langchain
langgraph
python-dotenv
pydantic
langchain_google_genai
langchain-ollama